Websocket update.
Refactors websocket commands to be sent to one websocket, if specified, or all websockets if not.  Also stores the most recent target, position, and connection status, and sends that data to newly connecting UIs, so if you open the UI after you've been out and about in the game, the UI will be up to date.